+++ src/main/scala/ai/tripl/arc/load/ORCLoad.scala
@@ -16,9 +16,22 @@ import ai.tripl.arc.util.EitherUtils._
 import ai.tripl.arc.util.ListenerUtils
 import ai.tripl.arc.util.Utils
 
-class ORCLoad extends PipelineStagePlugin {
+class ORCLoad extends PipelineStagePlugin with JupyterCompleter {
 
   val version = Utils.getFrameworkVersion
+
+  val snippet = """{
+    |  "type": "ORCLoad",
+    |  "name": "ORCLoad",
+    |  "environments": [
+    |    "production",
+    |    "test"
+    |  ],
+    |  "inputView": "inputView",
+    |  "outputURI": "hdfs://*.orc"
+    |}""".stripMargin
+
+  val documentationURI = new java.net.URI(s"${baseURI}/load/#orcload")
 
   def instantiate(index: Int, config: com.typesafe.config.Config)(implicit spark: SparkSession, logger: ai.tripl.arc.util.log.logger.Logger, arcContext: ARCContext): Either[List[ai.tripl.arc.config.Error.StageError], PipelineStage] = {
     import ai.tripl.arc.config.ConfigReader._
@@ -81,11 +94,12 @@ case class ORCLoadStage(
     authentication: Option[Authentication],
     saveMode: SaveMode,
     params: Map[String, String]
-  ) extends PipelineStage {
+  ) extends LoadPipelineStage {
 
   override def execute()(implicit spark: SparkSession, logger: ai.tripl.arc.util.log.logger.Logger, arcContext: ARCContext): Option[DataFrame] = {
     ORCLoadStage.execute(this)
   }
+
 }
 
 object ORCLoadStage {