+++ src/main/scala/ai/tripl/arc/transform/SimilarityJoinTransform.scala
@@ -21,9 +21,26 @@ import ai.tripl.arc.util.Utils
 import ai.tripl.arc.util.DetailException
 import ai.tripl.arc.util.EitherUtils._
 
-class SimilarityJoinTransform extends PipelineStagePlugin {
+class SimilarityJoinTransform extends PipelineStagePlugin with JupyterCompleter {
 
   val version = Utils.getFrameworkVersion
+
+  val snippet = """{
+    |  "type": "SimilarityJoinTransform",
+    |  "name": "SimilarityJoinTransform",
+    |  "environments": [
+    |    "production",
+    |    "test"
+    |  ],
+    |  "threshold": 0.75,
+    |  "leftView": "leftView",
+    |  "leftFields": [],
+    |  "rightView": "rightView",
+    |  "rightFields": [],
+    |  "outputView": "outputView"
+    |}""".stripMargin
+
+  val documentationURI = new java.net.URI(s"${baseURI}/transform/#similarityjointransform")
 
   def instantiate(index: Int, config: com.typesafe.config.Config)(implicit spark: SparkSession, logger: ai.tripl.arc.util.log.logger.Logger, arcContext: ARCContext): Either[List[ai.tripl.arc.config.Error.StageError], PipelineStage] = {
     import ai.tripl.arc.config.ConfigReader._
@@ -110,11 +127,12 @@ case class SimilarityJoinTransformStage(
     partitionBy: List[String],
     numPartitions: Option[Int],
     params: Map[String, String]
-  ) extends PipelineStage {
+  ) extends TransformPipelineStage {
 
   override def execute()(implicit spark: SparkSession, logger: ai.tripl.arc.util.log.logger.Logger, arcContext: ARCContext): Option[DataFrame] = {
     SimilarityJoinTransformStage.execute(this)
   }
+
 }
 
 object SimilarityJoinTransformStage {