+++ src/test/scala/ai/tripl/arc/transform/typing/BinaryTypingSuite.scala
@@ -61,7 +61,6 @@ class BinaryTypingSuite extends FunSuite with BeforeAndAfter {
         assert(err === TypingError("name", s"""Unable to convert '${value}' to binary using 'base64' decoding."""))
       }
       case (_,_) => {
-        ("Here")
         assert(false)
       }
     }